Move internal packages to internal path
Match repo structure used by other projects that I maintain.
