+++ Makefile
@@ -27,7 +27,7 @@ WHAT					= bridge
 
 # What package holds the "version" variable used in branding/version output?
 # VERSION_VAR_PKG			= $(shell go list -m)
-VERSION_VAR_PKG			=  $(shell go list -m)/config
+VERSION_VAR_PKG			=  $(shell go list -m)/internal/config
 
 OUTPUTDIR 				= release_assets
