+++ internal/matches/matches.go
@@ -19,9 +19,9 @@ import (
 	"strconv"
 	"text/tabwriter"
 
-	"github.com/atc0005/bridge/checksums"
-	"github.com/atc0005/bridge/paths"
-	"github.com/atc0005/bridge/units"
+	"github.com/atc0005/bridge/internal/checksums"
+	"github.com/atc0005/bridge/internal/paths"
+	"github.com/atc0005/bridge/internal/units"
 
 	"github.com/xuri/excelize/v2"
 )
